+++ src/components/Common/Tooltip.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import type { ICraftingInfo } from '@/data/interfaces';
 import { computed, ref, watch } from 'vue';
+import TooltipDescription from './TooltipDescription.vue';
 
 const props = defineProps<{
     targetElement: HTMLElement | null,
@@ -10,8 +11,9 @@ const props = defineProps<{
     description: string,
     effect: string | null,
     bonuses: { statName: string, value: number }[],
-    enhanced?: boolean,
-    enhancedBonus?: string,
+    enhancedDescription?: string,
+    enhancedEffects?: string,
+    enhancedBonuses: { statName: string, value: number }[],
 }>();
 
 // Dimensions in units of px
@@ -23,13 +25,15 @@ const mouseY = ref(0);
 const showTooltip = computed(() => !!props.targetElement && mouseIsHovering.value);
 const tooltipTop = computed(() => `${mouseY.value}px`);
 const tooltipLeft = computed(() => `${mouseX.value - tooltipWidth}px`);
-const showDescription = computed(() => !!props.description);
-const showBonuses = computed(() => props.bonuses && props.bonuses.length > 0);
 
 const requiredCraftLevel = computed(() => props.craftingInfo && props.craftingInfo.requiredLevel);
 const requiredCraftLevelEnhanced = computed(() => props.craftingInfo && props.craftingInfo.requiredLevelEnhanced);
 const requiredLevelText = computed(() => requiredCraftLevel.value ? `Lv${requiredCraftLevel.value}` : "");
 const requiredEnhancedLevelText = computed(() => requiredCraftLevelEnhanced.value ? `Lv${requiredCraftLevelEnhanced.value}` : "");
+
+const showEnhanced = computed(() => !!props.enhancedDescription || !!props.enhancedEffects || props.enhancedBonuses.length > 0);
+const descriptionEnhanced = computed(() => props.enhancedDescription || "");
+const effectsEnhanced = computed(() => props.enhancedEffects || "");
 
 watch(() => props.targetElement, (newValue, oldValue) => {
     if (oldValue) {
@@ -64,45 +68,36 @@ function onMouseEnter() {
 function onMouseLeave() {
     mouseIsHovering.value = false;
 }
-
-function textForStatValue(value: number) {
-    const prefix = value > 0 ? "+" : "";
-    return `${prefix}${value}`;
-}
 </script>
 
 <template>
     <div class="tooltip" :style="{ 'top': tooltipTop, 'left': tooltipLeft, }" v-show="showTooltip">
         <div class="header" v-if="!!props.title || !!props.subtitle">
             <h4 class="title" v-if="!!props.title">{{ props.title }}</h4>
-            <div class="subtitle" v-if="!!props.subtitle">
-                {{ props.subtitle }}
+            <div class="subheader" v-if="!!props.subtitle">
+                <span class="subtitle">{{ props.subtitle }}</span>
                 <span class="level" v-if="!!requiredCraftLevel">{{ requiredLevelText }}</span>
             </div>
         </div>
 
         <div class="body">
-            <div class="description" v-if="showDescription">{{ props.description }}</div>
-            <div class="effect" v-if="props.effect">{{ props.effect }}</div>
-
-            <div class="stats" v-if="showBonuses">
-                <ul class="stats-list">
-                    <li class="item" v-for="bonus of props.bonuses" :key="bonus.statName">
-                        <span class="name">{{ bonus.statName }}</span>
-                        <span
-                            class="value"
-                            :class="{ 'positive': bonus.value > 0, 'negative': bonus.value < 0, }"
-                        >{{ textForStatValue(bonus.value) }}</span>
-                    </li>
-                </ul>
+            <TooltipDescription
+                class="description"
+                :description="props.description"
+                :effect="props.effect"
+                :bonuses="props.bonuses"
+            ></TooltipDescription>
+            <div class="enhanced description" v-if="showEnhanced">
+                <div class="subheader">
+                    <span class="subtitle">Enhanced Benefits</span>
+                    <span class="level" v-if="!!requiredCraftLevelEnhanced">{{ requiredEnhancedLevelText }}</span>
+                </div>
+                <TooltipDescription
+                    :description="descriptionEnhanced"
+                    :effect="effectsEnhanced"
+                    :bonuses="props.enhancedBonuses"
+                ></TooltipDescription>
             </div>
-        </div>
-        <div class="enhanced" v-if="props.enhanced">
-            <div class="enhanced-title">
-                Enhanced Benefits
-                <span class="level" v-if="!!requiredCraftLevelEnhanced">{{ requiredEnhancedLevelText }}</span>
-            </div>
-            <div class="enhanced-bonus">{{ props.enhancedBonus }}</div>
         </div>
     </div>
 </template>
@@ -119,118 +114,71 @@ function textForStatValue(value: number) {
     display: flex;
     flex-direction: column;
 
-    background-color: black;
     color: #eee;
     line-height: 1.4;
     text-align: left;
 }
 
 .header {
-    border-top-width: 5px;
-    border-top-style: solid;
-    border-top-color: #eee;
+    border-top: 5px solid #eee;
     padding-top: 8px;
     padding-bottom: 8px;
     padding-left: 16px;
     padding-right: 16px;
+
+    background-color: rgba(11, 14, 17, 0.95);
 }
 
 .title {
     margin: 0;
     font-size: 20.8px;
     font-weight: 600;
-
     text-transform: uppercase;
 }
 
-.subtitle {
+.subheader {
     display: flex;
-    margin-top: 4px;
+    align-items: center;
     font-size: 14.4px;
     font-weight: 500;
 
-    text-transform: capitalize;
-}
+    .subtitle {
+        text-transform: capitalize;
 
-.level {
-    margin-left: auto;
-    text-transform: none;
+        .enhanced & {
+            text-transform: uppercase;
+            letter-spacing: 1px;
+            color: #ffce1f;
+        }
+    }
+
+    .level {
+        margin-left: auto;
+        font-size: 11.2px;
+    }
+
+    .enhanced & {
+        padding-top: 2px;
+        padding-bottom: 4px;
+        font-size: 11.2px;
+    }
 }
 
 .body {
     display: flex;
     flex-direction: column;
-    gap: 4px;
+    background-color: rgba(26, 26, 26, 0.9);
+}
 
-    margin: 0;
+.description {
     padding-top: 8px;
     padding-bottom: 8px;
     padding-left: 16px;
     padding-right: 16px;
+    background-color: transparent;
 
-    background-color: rgba(26, 26, 26, 0.9);
-    font-size: 12.8px;
-}
-
-.description {
-    font-family: neue-haas-grotesk-text, "Helvetica Neue", sans-serif;
-    font-weight: 300;
-    font-style: normal;
-    white-space: pre-wrap;
-}
-
-.effect {
-    font-family: neue-haas-grotesk-text, "Helvetica Neue", sans-serif;
-    font-size: 11.2px;
-    font-weight: 500;
-    white-space: pre-wrap;
-    color: #949494;
-}
-
-.stats {
-    margin-top: 8px;
-    padding-top: 8px;
-    border-top-width: 1px;
-    border-top-style: solid;
-    border-top-color: #fafafa;
-}
-
-.stats-list {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-}
-.stats-list .item {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-    margin-bottom: 4px;
-
-    line-height: 18.4px;
-    font-size: 16px;
-    font-family: neue-haas-grotesk-text, "Helvetica Neue", sans-serif;
-}
-.stats-list .name {
-    display: flex;
-    align-items: center;
-    padding-right: 16px;
-
-    font-size: 12.8px;
-    line-height: 12.8px;
-    text-align: right;
-}
-.stats-list .value {
-    display: flex;
-    align-items: center;
-    flex-basis: 60%;
-    font-size: 12.8px;
-
-    &.positive {
-        color: #68cc79;
-    }
-    &.negative {
-        color: #e25954;
+    &.enhanced {
+        padding-top: 4px;
     }
 }
 
@@ -241,6 +189,7 @@ function textForStatValue(value: number) {
     border-top-color: #ffce1f;
 
     &::before {
+        content: "";
         position: absolute;
         top: 0;
         left: 0;
@@ -248,23 +197,5 @@ function textForStatValue(value: number) {
         height: 30px;
         background: linear-gradient(0deg, rgba(183, 140, 37, 0), rgba(255, 206, 31, 0.1));
     }
-}
-.enhanced-title {
-    display: flex;
-    margin-bottom: 4px;
-    font-size: 11.2px;
-    font-family: neue-haas-grotesk-text, "Helvetica Neue", sans-serif;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    color: #ffce1f;
-}
-.enhanced-bonus {
-    margin: 0;
-    padding-bottom: 4px;
-
-    white-space: pre-wrap;
-    font-family: neue-haas-grotesk-text, "Helvetica Neue", sans-serif;
-    font-weight: 300;
-    font-style: normal;
 }
 </style>